+++ lib/lighthouseProvider/LighthouseDevice.dart
@@ -38,8 +38,7 @@ abstract class LighthouseDevice {
   ///
   /// Convert a device specific state byte to a global `LighthousePowerState`.
   ///
-  @protected
-  LighthousePowerState internalGetPowerStateFromByte(int byte);
+  LighthousePowerState powerStateFromByte(int byte);
 
   ///
   /// Get the update interval that this device supports.
@@ -75,7 +74,7 @@ abstract class LighthouseDevice {
 
   ///Get the power state of the device as a [LighthousePowerState] "enum".
   Stream<LighthousePowerState> get powerStateEnum => this.powerState.map((e) {
-        return internalGetPowerStateFromByte(e);
+        return powerStateFromByte(e);
       });
 
   BehaviorSubject<int> _powerState = BehaviorSubject.seeded(0xFF);