+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Close connection to discovering lighthouses on app close.
  - (Android) Hopefully fixed rare crash when closing app while scan is running.
  - Make sure no read is happening while writing.
+ - Added a dialog for changing the state when the state is unknown
 
 # Version 1.0.1+3 2020-08-31